MVC Format
Modified into MVC Format
