+++ client.py
@@ -3,10 +3,21 @@ from network import Handler, poll
 import sys
 from threading import Thread
 from time import sleep
+from view import display, printEasterEgg
+
+
+
+
+########## MODEL ##########
 
 running = True
 myname = None
 active = False
+
+
+
+
+########## CONTROLLER ##########
 
 class Client(Handler):
 
@@ -16,13 +27,14 @@ class Client(Handler):
 	def on_close(self):
 		global running
 		running = False
-		print "Goodbye! Press Enter to end program."
+		display("Goodbye! Press Enter to end program.")
 
 	def on_msg(self, msg):
 		global active
 		if msg == 'Connecting Now!':
 			active = True
-		print msg
+		display(msg)
+
 
 def chatSettings():
 	global myname
@@ -34,9 +46,9 @@ def chatSettings():
 	while (myname == ""):
 		myname = raw_input('Invalid. Please put a valid name: ')
 	if (personnel == "customer"):
-		print('1 - Complaint');
-		print('2 - Question');
-		print('3 - Other');
+		display('1 - Complaint');
+		display('2 - Question');
+		display('3 - Other');
 		myoption = raw_input('Please select an option: ');
 		while (myoption != "1") and (myoption != "2") and (myoption != "3"):
 			myoption = raw_input('Invalid. Please select an option: ')
@@ -61,21 +73,10 @@ thread.daemon = True  # die when the main thread dies
 thread.start()
 
 
-def printEasterEgg():
-	print('');
-	print('Easter egg!');
-	print('How we felt while making this project');
-	print('');
-	print('(╯°□°）╯︵ ┻━┻');
-	print('');
-	print('How we felt finishing this project');
-	print('');
-	print('┏━┓ ︵ /(^.^/)');
-	print('');
-
-
 
 while running:
+	global mytxt
+
 	mytxt = raw_input('')
 	if (not active):
 		client.do_send({'name': myname, 'txt': mytxt})
@@ -89,82 +90,3 @@ while running:
 			printEasterEgg()
 		else:
 			client.do_send({'name': myname, 'txt': mytxt})
-
-
-
-
-
-
-
-
-
-"""
-######## CONTROLLER ########
-
-userType = None; # Customer or Agent
-name = None;
-option = None;
-topic = None;
-
-
-print('Welcome to chat!');
-message = raw_input('Please enter your name: ');
-
-while (message != ':q'):
-	message = message.rstrip();
-	if (message == ':s'):
-		saveCopyOfChat();
-		if (name is None):
-			message = raw_input('Please enter your name: ');
-		elif (option is None):
-			print('1 - Complaint');
-			print('2 - Question');
-			print('3 - Other');
-			message = raw_input('Please select an option: ');
-		elif (topic is None):
-			message = raw_input('Please enter the topic: ');
-		else:
-			message = raw_input(name + ': ');
-	elif (message == ':e'):
-		printEasterEgg();
-		if (name is None):
-			message = raw_input('Please enter your name: ');
-		elif (option is None):
-			print('1 - Complaint');
-			print('2 - Question');
-			print('3 - Other');
-			message = raw_input('Please select an option: ');
-		elif (topic is None):
-			message = raw_input('Please enter the topic: ');
-		else:
-			message = raw_input(name + ': ');
-	else:
-		if (name is None):
-			name = message;
-			client.do_send(name + ' has joined the chat');
-			print('Hello ' + name + '!');
-			print('1 - Complaint');
-			print('2 - Question');
-			print('3 - Other');
-			message = raw_input('Please select an option: ');
-		elif (option is None):
-			if ((message == '1') or (message == '2') or (message == '3')):
-				option = message;
-				client.do_send(name + ' has selected option: ' + option);
-				print('You have selected option: ' + option);
-				message = raw_input('Please enter the topic: ');
-			else:
-				print('You have selected an invalid option: ' + message);
-				message = raw_input('Please try again: ');
-		elif (topic is None):
-			topic = message;
-			client.do_send(name + ' says the topic of the conversation is: ' + topic);
-			print('You have stated that the topic is: ' + topic);
-			print('You may start your conversation!');
-			message = raw_input(name + ': ');
-		else:
-			client.do_send(name + ': ' + message);
-			message = raw_input(name + ': ');
-client.do_send(name + ' has left the chat');
-sys.exit("Quitted chat");
-"""
