+++ server.py
@@ -1,5 +1,10 @@
 from network import Listener, Handler, poll
 import Queue
+from view import display
+
+
+
+########## MODEL ##########
 
 # Current Customer and Agent
 customerName = None
@@ -8,6 +13,11 @@ agentName = None
 agentHandler = None
 
 queue = Queue.Queue() # Queue of remaining Customers
+
+
+
+
+########## CONTROLLER ##########
 
 class MyHandler(Handler):
 
@@ -27,14 +37,14 @@ class MyHandler(Handler):
             if msg['personnel'] == 'agent':
                 agentName = msg['join']
                 agentHandler = self
-                print 'Agent ' + agentName + ' has connected successfully!'
+                display('Agent ' + agentName + ' has connected successfully!')
                 agentHandler.do_send('Welcome to the Chat System. Please wait to be connected')
                 agentHandler.do_send('Connecting Now!')
             elif msg['personnel'] == 'customer':
                 if (customerName is None) and (customerHandler is None):
                     customerName = msg['join']
                     customerHandler = self
-                    print 'Customer ' + customerName + ' has connected successfully!'
+                    display('Customer ' + customerName + ' has connected successfully!')
                     agentHandler.do_send('Customer ' + customerName + ' is being connected')
                     agentHandler.do_send(handleOption(msg['option']))
                     agentHandler.do_send('Topic is: ' + msg['topic'])
@@ -44,11 +54,11 @@ class MyHandler(Handler):
                     customerHandler.do_send('You are now connected to Agent ' + agentName)
                 else:
                     queue.put({'name': msg['join'], 'option': msg['option'], 'topic': msg['topic'], 'handler': self})
-                    print 'Customer ' + msg['join'] + ' is waiting in the queue.'
+                    display('Customer ' + msg['join'] + ' is waiting in the queue.')
                     self.do_send('Welcome to the Chat System. Please wait to be connected')
                     self.do_send('All available agents are busy. Please wait')
             else:
-                print 'Error'
+                display('Error')
         elif 'txt' in msg:
             if msg['name'] == agentName:
                 customerHandler.do_send(agentName + ": " + msg['txt'])
@@ -58,13 +68,13 @@ class MyHandler(Handler):
                 self.do_send('All available agents are busy. Please wait')
         elif 'special' in msg:
             if msg['special'] == "q":
-                print 'Customer ' + customerName + ' has quit the chat'
+                display('Customer ' + customerName + ' has quit the chat')
                 agentHandler.do_send('Customer ' + customerName + ' has quit the chat\n\n\n')
                 if (not queue.empty()):
                     nextCustomer = queue.get()
                     customerName = nextCustomer['name']
                     customerHandler = nextCustomer['handler']
-                    print 'Customer ' + customerName + ' has connected successfully'
+                    display('Customer ' + customerName + ' has connected successfully')
                     agentHandler.do_send('Customer ' + customerName + ' is being connected')
                     agentHandler.do_send(handleOption(nextCustomer['option']))
                     agentHandler.do_send('Topic is: ' + nextCustomer['topic'])
