+++ view.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+
+########## VIEW ##########
+
+def display(message):
+	print message
+
+def printEasterEgg():
+	print('');
+	print('Easter egg!');
+	print('How we felt while making this project');
+	print('');
+	print('(╯°□°）╯︵ ┻━┻');
+	print('');
+	print('How we felt finishing this project');
+	print('');
+	print('┏━┓ ︵ /(^.^/)');
+	print('');
